requirements.txt
decorator==4.0.10
oauthlib==2.0.1
olefile==0.44
Pillow==4.0.0
praw==4.2.0
prawcore==0.6.0
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
update-checker==0.16
